Changed floor to floorf for VC 7 build.

+++ actors/Actor.cpp
@@ -644,7 +644,7 @@ Obj *Actor::inventory_new_object(uint16 obj_n, uint32 qty, uint8 quality)
  
  if(obj_n > 1024)
    {
-    frame_n = (uint8)floor(obj_n / 1024);
+    frame_n = (uint8)floorf(obj_n / 1024);
 	obj_n -= frame_n * 1024;
    }
